+++ protocols/identify/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `ConnectionId` in `Event`.
   See [PR 4981](https://github.com/libp2p/rust-libp2p/pull/4981).
+- Add `hide_listen_addrs` option to prevent leaking (local) listen addresses.
 
 ## 0.44.2